% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffPeaks.R
\name{DiffPeaks}
\alias{DiffPeaks}
\title{Function to perform differential analysis on peaks using mixed-effects logistic regression}
\usage{
DiffPeaks(
  metadata,
  provided.matrix,
  log.OR.threshold = 0.1,
  min.pct = 0.01,
  subset = NULL,
  sample.column,
  cluster.column,
  selected.clusters,
  treatment.column,
  treatment.levels,
  fixed.effect,
  random.effect,
  ncores = 1
)
}
\arguments{
\item{metadata}{Data frame containing the required metadata}

\item{provided.matrix}{Matrix or sparse matrix class to use as input. Overrides useMatrix}

\item{log.OR.threshold}{Limit testing to peaks which show at least an odds-ratio of X (log-scale) between the two groups of cells.}

\item{min.pct}{Only test peaks that are detected in a minimum fraction of min.pct cells in either of the two populations}

\item{subset}{Character vector containing names of features to subset for differential analysis}

\item{sample.column}{Character vector of length one indicating the column containing the sample identity}

\item{cluster.column}{Character vector of length one indicating the column containing the cluster identity}

\item{selected.clusters}{Character vector indicating the clusters to be included in the analysis}

\item{treatment.column}{Character vector of length one indicating the column containing the treatments}

\item{treatment.levels}{Character vector of length two indicating which treatments to compare}

\item{fixed.effect}{Character vector indicating the predictor variables in the model that are assumed to have the same effect on the response variable (treatment) for all the units (peaks) under study. Each variable should be a column name in the metadata dataframe.}

\item{random.effect}{Character vector indicating the predictor variables in the model that are assumed to have different effects on the response variable (treatment) for different units (peaks) under study. Each variable should be a column name in the metadata dataframe.}

\item{ncores}{Integer indicating how many cores to use for parallel computing}
}
\value{
Data frame containing the results of the differential analysis
}
\description{
Function to perform differential analysis on peaks using mixed-effects logistic regression
}
