% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeFastqs.R
\name{MergeFastqs}
\alias{MergeFastqs}
\title{Merge multi-sample Gotcha fastqs}
\usage{
MergeFastqs(path = "gotcha_fastqs/", out.dir = "merged")
}
\arguments{
\item{path}{Path to Gotcha fastqs}

\item{out.dir}{Name of new directory to save merged fastqs}
}
\value{
New directory with merged fastqs
}
\description{
Merge multi-sample Gotcha fastqs
}
